CorrectBot: implement @use_snippets decorator
Correction functions can now easily define if they need to be run on the
snippets (decorate the function with @use_snippets) or if they can just
be run on the full translation unit (default). For this, the correction
functions in CorrectorBase need to take TranslationUnits as arguments.
Improved some rules to reduce miscorrections and added tests.