Save (English) language name in class LanguageInfo
The English name of a language can be queried with
fortraininglib.get_language_name(). Within the resourcesbot we currently
only need that in WriteList but as more post-processing plugins will need
that information as well, it's better to directly put it into LanguageInfo
than doing this low-level query several times.