Fill changelog when reading from cache
Bugfix that caused Exception in WriteReport when resources_bot.py was
called with --read-from-cache.
Now changelog gets correctly filled also when reading from cache.
WriteReport is more robust to not throw exception in such a case.
Added a test case for this: A first test of ResourcesBot.run()
It's not yet covering all code paths but the harder part of simulating
all the other code parts involved is done :)