Refactor fortraininglib into a class
New class ForTrainingLib.
This gets rid of the ugly hard-coded BASEURL="https://www.4training.net"
constant that was in fortraininglib before. Instead we can read the
mediawiki system we want to connect to from the config.ini
Now https://www.4training.net is hard-coded on several places but that
makes it at least transparent where there is still some work to be done.