Refactor sorting of TranslationUnits
The previous implementation was incorrect and did only re-order in some
places and not all where it was necessary. As this is no normal sorting,
we can't use the built-in sort() function but need to compare each unit
with each other unit.
Clean up units in TranslateODT before we do any sorting.