CorrectBot says if nothing was saved; NOEDITSECTION
Did CorrectBot write a report? If no, then write that in the output
together with a reason.
TODO still no tests for CorrectBot.run()
Adding __NOEDITSECTION__ magic word to correctbot and language summary
report to avoid "edit" buttons (which are useless and might confuse)
Add strip() for comparisons whether a mediawiki page needs to be saved
or whether the content is already up-to-date