CorrectBot: Recognize warnings in corrector functions
If a corrector function logs a warning, it's now correctly recognized in
the CorrectionResult and mentioned in the report written. We do this by
catching potential logger messages in the CorrectorBase class.
Closes #70
Make check_for_typos.py functional again
Fixing FrenchCorrector.correct_quotation_marks() - it was only
recognizing special quotation marks if they happened to be at the
beginning (and unfortunately the test case had them in the beginning so
everything seemed fine)