Resolve import issues
Move scripts to base folder in this repository so that we don't need to
mess around with PYTHONPATH anymore. We consistently use absolute
imports.
Split translateodt.py in three parts: The one in the base folder only
contains argument handling, the main code is in
pywikitools/translateodt/__init__.py
Close #41