Reorder CorrectBot report contents
Change the order of the CorrectBot reports to first show all warnings as
they definitely need attention. Afterwards show suggestions as they also
would need to be looked at. In the end all corrections follow - can be
last as nothing needs to be done, that section is just for information,
Closes #83