GenerateOdtBot: Log in if we're not yet logged in
In case generateodtbot.py fails with
pywikibot.exceptions.UserRightsError: You don't have permission to send
mail, this is because the mediawiki user is currently logged in.
However, in that case the command needs to be run by the command line in
order to be able to enter the password. After that the user is logged in
for some time (until the login cookie saved in pywikibot.lwp expires)
and it can be invoked also in other ways (not from command line)